pyproject.toml
[project]
name = "librarian"
version = "0.1.0"
description = "Command line tool and library for manage your vector store"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.0",
    "langchain-community>=0.3.31",
    "langchain-core>=0.3.78",
    "langchain-huggingface>=0.3.1",
    "langchain-qdrant>=0.2.1",
    "langchain-text-splitters>=0.3.11",
    "langchain-unstructured>=0.1.5",
    "pypandoc-binary>=1.15",
    "pypdf>=6.1.1",
    "qdrant-client>=1.15.1",
    "rich>=13.0.0",
    "sentence-transformers>=5.1.1",
    "sqlalchemy>=2.0.43",
    "tiktoken>=0.12.0",
    "torch>=2.8.0",
    "tqdm>=4.66.0",
    "unstructured[epub,md]>=0.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["librarian*"]

[project.scripts]
librarian = "librarian.cli.librarian_cli:librarian"
